Fix a compilation warning so I can get it out of my tree which has a large unrelated patch in it.
git-svn-id: http://svn.mythtv.org/svn/trunk@6564 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/NuppelVideoPlayer.cpp
@@ -4262,7 +4262,7 @@ int NuppelVideoPlayer::calcSliderPos(QString &desc)
         
         
         if (rPos)
-            return  (rPos / tPos) * 1000.0;
+            return (int)((rPos / tPos) * 1000.0);
         else
             return 0;
     }
